--- src/analyze_csv.py
@@ -1,5 +1,0 @@
-def analyze():
-    pass
-
-if __name__ == '__main__':
-    analyze()
